+++ src/domain/FoodStorage.ts
@@ -2,7 +2,7 @@ import { FoodType } from "../types/component/FoodItemType";
 import {
   DeleteStorageFoodListType,
   UpdateStorageFoodListType,
-} from "../types/domain/FoodStorageHandlerType";
+} from "../types/domain/FoodStorageType";
 
 // CRUD - read
 export function readStorageFoodList() {
