+++ src/types/domain/FavoriteServiceType.ts
