+++ src/types/domain/FoodStorageType.ts
