+++ src/util/errorDisplay.ts
@@ -1,5 +1,5 @@
 import { Alert } from "../component/layout/Alert.js";
-import { AlertErrorType } from "../types/util/AlertErrorType.js";
+import { AlertErrorType } from "../types/util/ErrorDisplayType.js";
 
 export function alertError({ error }: AlertErrorType) {
   if (!document.querySelector(".alert")) {
